fix client might be blocked
If a server loses contact with its leader, it will update its term for election. Since other servers are all right, the election will not succeed. But if the client learns about the new term and updates its term to it, it will never get the true leader. Now we will reset client's term only when the server has a leader id.